BUG: only purge output repos

+++ astrocats/catalog/gitter.py
@@ -188,7 +188,7 @@ def git_clone_all_repos(cat):
             log.debug("Cloning directory...")
             clone(repo, cat.log, depth=max(cat.args.clone_depth, 1))
 
-        if cat.args.purge_outputs:
+        if cat.args.purge_outputs and repo in out_repos:
             for fil in glob(os.path.join(repo, '*.json')):
                 os.remove(fil)
